Refactor 2.1-2.4
